feat: add shared validators module
