+++ conversation_service/agents/base_agent.py
@@ -34,9 +34,10 @@ from openai import AsyncOpenAI
 
 # Local imports
 from ..models.agent_models import AgentConfig, AgentResponse
-from ..models.core_models import HarenaValidators, FinancialEntity
+from ..models.core_models import FinancialEntity
 from ..core.cache_manager import CacheManager
 from ..core.metrics_collector import MetricsCollector
+from ..core.validators import HarenaValidators
 from ..utils.logging import get_structured_logger
 
 __all__ = ["BaseFinancialAgent", "AgentPerformanceTracker", "PromptOptimizer"]
